timur: merged 2 servers and dbs
