+++ package.json
@@ -22,8 +22,7 @@
     "lint:check": "eslint \"src/**/*.{ts,tsx}\"",
     "lint:fix": "eslint --fix \"src/**/*.{ts,tsx}\"",
     "analyze": "npm run build:dev && webpack-bundle-analyzer --port 4200 dist/stats.json",
-    "prettier": "prettier --write src",
-    "settings:build": "tsc --project settings/tsconfig.json"
+    "prettier": "prettier --write src"
   },
   "husky": {
     "hooks": {
@@ -40,6 +39,9 @@
     "@loadable/component": "^5.15.0",
     "@loadable/server": "^5.15.0",
     "babel-jest": "^27.0.2",
+    "body-parser": "^1.19.0",
+    "cookie-parser": "^1.4.5",
+    "cors": "^2.8.5",
     "connected-react-router": "^6.9.1",
     "express": "^4.17.1",
     "express-query-parser": "^1.2.0",
@@ -60,7 +62,6 @@
     "sequelize-typescript": "^2.1.0"
   },
   "devDependencies": {
-    "@types/cookie-parser": "^1.4.2",
     "@babel/core": "^7.13.16",
     "@babel/plugin-proposal-decorators": "^7.14.5",
     "@babel/plugin-transform-runtime": "^7.14.5",
@@ -69,13 +70,16 @@
     "@babel/preset-typescript": "^7.13.0",
     "@loadable/babel-plugin": "^5.13.2",
     "@loadable/webpack-plugin": "^5.15.0",
+    "@types/body-parser": "^1.19.1",
+    "@types/cookie-parser": "^1.4.2",
+    "@types/cors": "^2.8.12",
     "@types/express": "^4.17.12",
     "@types/jest": "^26.0.23",
-    "@types/pg": "^8.6.0",
     "@types/loadable__component": "^5.13.4",
     "@types/loadable__server": "^5.12.6",
     "@types/loadable__webpack-plugin": "^5.7.3",
     "@types/node": "^15.12.4",
+    "@types/pg": "^8.6.0",
     "@types/react": "^17.0.4",
     "@types/react-dom": "^17.0.3",
     "@types/react-redux": "^7.1.16",
