--- settings/package.json
@@ -1,87 +1,0 @@
-{
-  "name": "bomber-game",
-  "version": "1.0.0",
-  "description": "Командный проект: игра Бомбер.",
-  "main": "index.js",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/odessa-yandex-praktikum/crazy-bomber.git"
-  },
-  "author": "odessa-team",
-  "license": "ISC",
-  "homepage": "https://github.com/odessa-yandex-praktikum/crazy-bomber#readme",
-  "bugs": {
-    "url": "https://github.com/odessa-yandex-praktikum/crazy-bomber/issues"
-  },
-  "scripts": {
-    "settings:build": "tsc",
-    "settings:serve": "ts-node ./src/server.ts"
-  },
-  "dependencies": {
-    "@types/cookie-parser": "^1.4.2",
-    "cookie-parser": "^1.4.5",
-    "cors": "^2.8.5",
-    "express-query-parser": "^1.2.0",
-    "express-router": "0.0.1",
-    "identity-obj-proxy": "^3.0.0",
-    "parser": "^0.1.4",
-    "pg": "^8.6.0",
-    "pg-hstore": "^2.3.4",
-    "query": "^0.2.0",
-    "router": "^1.3.5",
-    "search-query-parser": "^1.6.0",
-    "sequelize": "^6.6.4",
-    "sequelize-typescript": "^2.1.0"
-  },
-  "devDependencies": {
-    "@babel/core": "^7.13.16",
-    "@babel/plugin-transform-runtime": "^7.14.5",
-    "@babel/preset-env": "^7.13.15",
-    "@babel/preset-react": "^7.13.13",
-    "@babel/preset-typescript": "^7.13.0",
-    "@types/body-parser": "^1.19.1",
-    "@types/express": "^4.17.12",
-    "@types/jest": "^26.0.23",
-    "@types/pg": "^8.6.0",
-    "@types/react": "^17.0.4",
-    "@types/react-dom": "^17.0.3",
-    "@types/react-redux": "^7.1.16",
-    "@types/react-router-dom": "^5.1.7",
-    "@types/redux": "^3.6.0",
-    "@types/redux-thunk": "^2.1.0",
-    "@typescript-eslint/eslint-plugin": "^4.22.0",
-    "@typescript-eslint/parser": "^4.22.0",
-    "@wojtekmaj/enzyme-adapter-react-17": "^0.6.2",
-    "babel-loader": "^8.2.2",
-    "babel-plugin-transform-class-properties": "^6.24.1",
-    "babel-polyfill": "^6.26.0",
-    "body-parser": "^1.19.0",
-    "css-loader": "^5.2.4",
-    "enzyme": "^3.11.0",
-    "eslint": "^7.25.0",
-    "eslint-config-airbnb-base": "^14.2.1",
-    "eslint-config-prettier": "^8.3.0",
-    "eslint-plugin-import-helpers": "^1.1.0",
-    "eslint-plugin-jest": "^24.3.6",
-    "eslint-plugin-prettier": "^3.4.0",
-    "eslint-plugin-react": "^7.23.2",
-    "express": "^4.17.1",
-    "file-loader": "^6.2.0",
-    "html-webpack-plugin": "^5.3.1",
-    "husky": "^6.0.0",
-    "jest": "^27.0.4",
-    "lint-staged": "^10.5.4",
-    "mini-css-extract-plugin": "^1.5.1",
-    "prettier": "^2.2.1",
-    "style-loader": "^2.0.0",
-    "ts-node": "^10.0.0",
-    "tslib": "^2.2.0",
-    "typescript": "^4.2.4",
-    "url-loader": "^4.1.1",
-    "webpack": "^5.36.1",
-    "webpack-bundle-analyzer": "^4.4.1",
-    "webpack-cli": "^4.6.0",
-    "webpack-dev-server": "^3.11.2",
-    "webpack-merge": "^5.7.3"
-  }
-}
