+++ src/server/server.ts
@@ -1,5 +1,16 @@
+import bodyParser from 'body-parser';
+import cookieParser from 'cookie-parser';
+import cors from 'cors';
 import * as express from 'express';
+import {Router} from 'express';
 import * as path from 'path';
+import {
+    getThemeHandler,
+    getUserTheme,
+    updateThemeHandler,
+    writeNewTheme,
+    writeThemeHandler,
+} from './controllers/theme';
 import {sequelize} from './db/db';
 import {getWebpackMiddlewares} from './middlewares';
 
@@ -13,7 +24,15 @@ const mode: 'development' | 'production' =
 console.log('[server started in mode]: ', mode);
 
 const app = express();
+const router: Router = Router();
 const PORT = 3000;
+
+app.use(
+    cors({
+        origin: 'http://localhost:3000',
+        credentials: true,
+    })
+);
 
 /**
  * Отдаём статику приложения.
@@ -24,6 +43,17 @@ app.use(express.static(path.posix.resolve('dist')));
  * На все get запросы запускаем сначала middleware dev server, а потом middleware рендеринга приложения.
  */
 app.get('/*', getWebpackMiddlewares(mode));
+
+app.use(bodyParser.urlencoded({extended: false}));
+app.use(bodyParser.json());
+
+app.get('/get-theme', getThemeHandler);
+app.post('/write-user-theme', writeThemeHandler);
+app.post('/update-user-theme', updateThemeHandler);
+app.get('/get-user-theme', getUserTheme);
+app.get('/write-new-theme', writeNewTheme);
+
+app.disable('x-powered-by').enable('trust proxy').use(cookieParser()).use(router);
 
 /**
  * Запуск приложения.
